% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_partial_dependence.R
\name{estimate_partial_dependence}
\alias{estimate_partial_dependence}
\title{Estimate partial dependence of model predictions to variables}
\usage{
estimate_partial_dependence(modelPath, variables)
}
\arguments{
\item{modelPath}{the path of the RData file where the model is saved}

\item{variables}{character vector. The names of the variables for which the
partial independence should be evaluated.}
}
\value{
a list with two elements: \code{data} the estimated partial dependence
values and \code{plot} a ggplot object representing the trends of the
relationships between model predictions and the values of the variable of
interest.
}
\description{
This functions evaluates and represents the partial dependence of a model
prediction to changes in individual variable values.
}
\details{
The function estimates (using \link[=mlr]{generatePartialDependenceData}) how the
model stored in \code{modelPath} is affected by changes in one of several
\code{variables} values. The partial dependence of the model for a given variable
is calculated by setting all the other variables to their mean values and by
exploring the changes in model prediction by setting the values of the
variable of interest over a range. If, during the construction of the DT unit
using \link{build_DT}, several models were trained (i.e. \code{nIter} larger than 1),
the partial dependence is calculated and represented for each of the
individual model.

Once the partial dependence values calculated, a plot is produced
representing the changes in model predictions over the range of the variable
values. A \link[=stats]{loess} smoothed curve is added to represent the tendencies.
The distribution of the values of the variable of interest in the training
data set is also represented.
}
\seealso{
\link[=mlr]{generatePartialDependenceData}
}
