% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_variable_importance.R
\name{estimate_variable_importance}
\alias{estimate_variable_importance}
\title{Estimate variable importance}
\usage{
estimate_variable_importance(modelPath, methods = c("anova.test", "auc",
  "chi.squared", "gain.ratio", "information.gain", "kruskal.test",
  "ranger.impurity", "ranger.permutation"), nVarToPlot = 20,
  nIter = 10, nCores = 1L)
}
\arguments{
\item{modelPath}{the path of the RData file where the model is saved}

\item{methods}{character vector. The metric(s) used to estimate variable
importance. The available choices are: \code{anova.test}, \code{auc}, \code{chi.squared}
(package \link{FSelector}), \code{gain.ratio} (\link{FSelector}), \code{information.gain}
(\link{FSelector}), \code{kruskal.test}, \code{ranger.impurity} (\link{ranger}) and/or
\code{ranger.permutation} (\link{ranger})}

\item{nVarToPlot}{numeric. The number of most important variables to
graphically represent}

\item{nIter}{integer. If \code{ranger.impurity} or \code{ranger.permutation} is used as
importance metrics, the number of times the estimate is repeated.}

\item{nCores}{integer.If \code{ranger.impurity} or \code{ranger.permutation} is used as
importance metrics and \code{nIter} larger than 1, the number of CPUs used to
perform the computations.}
}
\value{
a list with two elements: \code{varImp}: the table with the importance
measure(s) for all variables and \code{varImpPlot} a ggplot object representing
the importance of the most important variables.
}
\description{
This function allows to estimate the importance of individual variables in a
model unit of a diagnostic tool.
}
\details{
This functions estimates the variable importance of all the variables
included in the investigated model using the importance metric(s) specified
in the \code{method} argument. In this regard, the function is a wrapper around
the function \link[=mlr]{generateFilterValuesData} from the \code{mlr} package with the
possibility to run multiple iterations for the metrics \code{ranger.impurity} and
\code{ranger.permutation} potentially in parallel (using \code{nCores} larger than 1).

The second step performed by this function corresponds to the production of a
plot representing the importance of the most important variables. The
selection of the metrics is performed by ranking the importance metric values
and the number of variables to be represented is controlled by the argument
\code{nVarToPlot}. If several importance metrics are used, the selection is made
on the average rank of the variables over the different metrics.
}
\seealso{
\link[=mlr]{generateFilterValuesData}
}
