% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DT.R
\name{plot_DT}
\alias{plot_DT}
\title{Plot the predictions of a DT model}
\usage{
plot_DT(plot.data, axis.labels = colnames(plot.data)[-1],
  grid.line.width = 0.5, gridline.min.colour = "grey",
  gridline.mid.colour = "#007A87", gridline.max.colour = "grey",
  grid.label.size = 7, axis.label.offset = 1.15, axis.label.size = 8,
  axis.line.colour = "grey", group.line.width = 1.5,
  group.point.size = 6, background.circle.colour = "#D7D6D1",
  background.circle.transparency = 0.2, plot.legend = if
  (nrow(plot.data) > 1) TRUE else FALSE, legend.title = "",
  plot.title = "", legend.text.size = grid.label.size, byRow = FALSE)
}
\arguments{
\item{plot.data}{dataframe comprising samples in rows and DT model
predictions in columns (as returned by \link{run_DT})}

\item{axis.labels}{names of axis labels if other than column names supplied
via plot.data}

\item{grid.line.width, gridline.min.colour, gridline.mid.colour, gridline.max.colour}{controls for the appearance of some aspects the 'minimum', 'average' and
'maximum' grid lines.}

\item{axis.label.offset}{vertical displacement of axis labels from maximum
grid line, measured relative to circle diameter}

\item{axis.label.size, grid.label.size, legend.text.size}{text size}

\item{axis.line.colour}{colour of the axes}

\item{group.line.width, group.point.size}{controls for the appearance of some
aspects of the sample lines}

\item{background.circle.colour, background.circle.transparency}{controls for
the appearance of background circle}

\item{plot.legend}{logical, does the legend be printed or not}

\item{legend.title, plot.title}{titles}

\item{byRow}{logical, should one radar plot be returned for each sample ID}
}
\description{
This function is a wrapper around the
\href{https://github.com/ricardo-bion/ggradar/blob/master/R/ggradar.R}{ggradar}
function and produces radar plot for the DT model predictions.
}
\references{
script adapted from
https://github.com/ricardo-bion/ggradar/blob/master/R/ggradar.R
}
