% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DT.R
\name{run_DT}
\alias{run_DT}
\title{Run a Diagnostic Tool}
\usage{
run_DT(pathDT, newdata)
}
\arguments{
\item{pathDT}{character string, the path where the built models will be saved}

\item{newdata}{a data frame with samples in rows and biological metrics in
columns}
}
\value{
a data frame with the samples from \code{newdata} in rows and the
corresponding predictions for the different pressures included in the DT
model in columns.
}
\description{
This function run a diagnostic tool DT model built using the
\link{build_DT} function on new data.
}
\details{
The function takes the path where the DT model has been saved and the new
data as inputs. Then, the averaged predictions for each of the DT unit (unit
of a DT model corresponding to a given pressure) are computed and the
predictions of the different DT units are gathered.
}
\seealso{
\link{build_DT}
}
